requirements.txt
pandas
python-dotenv
google-generativeai
streamlit
faiss-cpu 
sentence-transformers 
python-dotenv
huggingface_hub
